move MaterialDispoDataItem to cucumber tests
