+++ backend/de.metas.cucumber/src/test/java/de/metas/cucumber/stepdefs/material/dispo/MaterialDispoDataItem_StepDefData.java
@@ -25,7 +25,6 @@ package de.metas.cucumber.stepdefs.material.dispo;
 import de.metas.cucumber.stepdefs.StepDefData;
 import de.metas.cucumber.stepdefs.StepDefDataGetIdAware;
 import de.metas.material.dispo.commons.candidate.CandidateId;
-import de.metas.material.dispo.commons.candidate.MaterialDispoDataItem;
 
 /**
  * Having a dedicated class to help the IOC-framework injecting the right instances, if a step-def needs more than one.
